# ui/static/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Scanopy API",
    "description": "Network topology discovery and visualization API",
    "license": {
      "name": "Dual (AGPL3.0, Commercial License Available)"
    },
    "version": "0.12.5"
  },
  "paths": {
...
    "/api/discovery": {
      "get": {
        "tags": [
          "discoveries"
        ],
        "summary": "List all discoveries",
        "operationId": "list_discoveries",
        "parameters": [
          {
            "name": "network_id",
            "in": "query",
            "description": "Filter by network ID",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ],
              "format": "uuid"
            }
          },
          {
            "name": "daemon_id",
            "in": "query",
            "description": "Filter by daemon ID",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ],
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of discoveries",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Discovery"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "session": []
          }
        ]
      },
      "post": {
        "tags": [
          "discoveries"
        ],
        "summary": "Create a new discovery",
        "operationId": "create_discovery",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Discovery"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Discovery created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Discovery"
                }
              }
            }
          },
          "400": {
            "description": "Invalid subnet network"
          }
        },
        "security": [
          {
            "session": []
          }
        ]
      }
    },
...
    "/api/discovery/bulk-delete": {
      "post": {
        "tags": [
          "discoveries"
        ],
        "summary": "Bulk delete discoveries",
        "operationId": "bulk_delete_discoveries",
        "requestBody": {
          "description": "Array of discoveries IDs to delete",
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Discoverys deleted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BulkDeleteResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "session": []
          }
        ]
      }
    },
...
    "/api/discovery/{id}": {
      "get": {
        "tags": [
          "discoveries"
        ],
        "summary": "Get discovery by ID",
        "operationId": "get_discovery_by_id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Discovery ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Discovery found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Discovery"
                }
              }
            }
          },
          "404": {
            "description": "Discovery not found"
          }
        },
        "security": [
          {
            "session": []
          }
        ]
      },
      "put": {
        "tags": [
          "discoveries"
        ],
        "summary": "Update discovery",
        "operationId": "update_discovery",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Discovery ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Discovery"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Discovery updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Discovery"
                }
              }
            }
          },
          "404": {
            "description": "Discovery not found"
          }
        },
        "security": [
          {
            "session": []
          }
        ]
      },
      "delete": {
        "tags": [
          "discoveries"
        ],
        "summary": "Delete discovery",
        "operationId": "delete_discovery",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Discovery ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Discovery deleted"
          },
          "404": {
            "description": "Discovery not found"
          }
        },
        "security": [
          {
            "session": []
          }
        ]
      }
    },
...
    "/api/hosts": {
      "get": {
        "tags": [
          "hosts"
        ],
        "summary": "List all hosts",
        "description": "Returns all hosts the authenticated user has access to, with their\ninterfaces, ports, and services included.",
        "operationId": "get_all_hosts",
        "responses": {
          "200": {
            "description": "List of hosts with their children",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/HostResponse"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "session": []
          }
        ]
      },
      "post": {
        "tags": [
          "hosts"
        ],
        "summary": "Create a new host",
        "description": "Creates a host with optional interfaces, ports, and services.\nThe `source` field is automatically set to `Manual`.\nIDs for the host and all children are generated server-side.",
        "operationId": "create_host",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateHostRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Host created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HostResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request - network not found or subnet mismatch"
          }
        },
        "security": [
          {
            "session": []
          }
        ]
      }
    },
    "/api/hosts/bulk-delete": {
      "post": {
        "tags": [
          "hosts"
        ],
        "summary": "Bulk delete hosts",
        "description": "Deletes multiple hosts in a single request. The request body should be\nan array of host IDs to delete. Fails if any host has an associated daemon.",
        "operationId": "bulk_delete_hosts",
        "requestBody": {
          "description": "Array of host IDs to delete",
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Hosts deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BulkDeleteResponse"
                }
              }
            }
          },
          "409": {
            "description": "One or more hosts has an associated daemon - delete daemons first"
          }
        },
        "security": [
          {
            "session": []
          }
        ]
      }
    },
    "/api/hosts/{destination_host}/consolidate/{other_host}": {
      "put": {
        "tags": [
          "hosts"
        ],
        "summary": "Consolidate hosts",
        "description": "Merges all interfaces, ports, and services from `other_host` into\n`destination_host`, then deletes `other_host`. Both hosts must be\non the same network.",
        "operationId": "consolidate_hosts",
        "parameters": [
          {
            "name": "destination_host",
            "in": "path",
            "description": "Destination host ID - will receive all children",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "other_host",
            "in": "path",
            "description": "Host to merge into destination - will be deleted",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Hosts consolidated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HostResponse"
                }
              }
            }
          },
          "400": {
            "description": "Hosts are on different networks"
          },
          "404": {
            "description": "One or both hosts not found"
          }
        },
        "security": [
          {
            "session": []
          }
        ]
      }
    },
    "/api/hosts/{id}": {
      "get": {
        "tags": [
          "hosts"
        ],
        "summary": "Get a host by ID",
        "description": "Returns a single host with its interfaces, ports, and services.",
        "operationId": "get_host_by_id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Host ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Host found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HostResponse"
                }
              }
            }
          },
          "404": {
            "description": "Host not found"
          }
        },
        "security": [
          {
            "session": []
          }
        ]
      },
      "put": {
        "tags": [
          "hosts"
        ],
        "summary": "Update a host",
        "description": "Updates host properties. Children (interfaces, ports, services)\nare managed via their own endpoints.",
        "operationId": "update_host",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Host ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateHostRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Host updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HostResponse"
                }
              }
            }
          },
          "404": {
            "description": "Host not found"
          }
        },
        "security": [
          {
            "session": []
          }
        ]
      },
      "delete": {
        "tags": [
          "hosts"
        ],
        "summary": "Delete a host",
        "description": "Prevents deletion if the host has a daemon associated with it",
        "operationId": "delete_host",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Host ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Host deleted"
          },
          "404": {
            "description": "Host not found"
          },
          "409": {
            "description": "Host has associated daemon"
          }
        },
        "security": [
          {
            "session": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BulkDeleteResponse": {
        "type": "object",
        "required": [
          "deleted_count",
          "requested_count"
        ],
        "properties": {
          "deleted_count": {
            "type": "integer",
            "minimum": 0
          },
          "requested_count": {
            "type": "integer",
            "minimum": 0
          }
        }
      },
      "CreateHostRequest": {
        "type": "object",
        "description": "Request type for creating a host with its associated interfaces and ports.\nServer assigns `host_id`, `network_id`, and `source` to all children.\nSource is automatically set based on how the entity was created (API vs UI).\n\nNote: Services are created separately via `POST /api/services` after the host exists,\nas service bindings require the real IDs of the interfaces/ports to reference.",
        "required": [
          "name",
          "network_id"
        ],
        "properties": {
          "description": {
            "type": [
              "string",
              "null"
            ]
          },
          "hidden": {
            "type": "boolean"
          },
          "hostname": {
            "type": [
              "string",
              "null"
            ]
          },
          "interfaces": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CreateInterfaceInput"
            }
          },
          "name": {
            "type": "string"
          },
          "network_id": {
            "type": "string",
            "format": "uuid"
          },
          "ports": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CreatePortInput"
            }
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            }
          },
          "virtualization": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/HostVirtualization"
              }
            ]
          }
        },
        "example": {
          "description": "Primary web server",
          "hidden": false,
          "hostname": "web-server-01.local",
          "interfaces": [
            {
              "ip_address": "192.168.1.100",
              "mac_address": "DE:AD:BE:EF:12:34",
              "name": "eth0",
              "subnet_id": "550e8400-e29b-41d4-a716-446655440004"
            }
          ],
          "name": "web-server-01",
          "network_id": "550e8400-e29b-41d4-a716-446655440002",
          "ports": [
            {
              "number": 80,
              "protocol": "Tcp"
            }
          ],
          "tags": [],
          "virtualization": null
        }
      },
...
      "Discovery": {
        "allOf": [
          {
            "$ref": "#/components/schemas/DiscoveryBase"
          },
          {
            "type": "object",
            "properties": {
              "created_at": {
                "type": "string",
                "format": "date-time",
                "readOnly": true
              },
              "id": {
                "type": "string",
                "format": "uuid",
                "readOnly": true
              },
              "updated_at": {
                "type": "string",
                "format": "date-time",
                "readOnly": true
              }
            }
          }
        ]
      },
...
      "HostResponse": {
        "type": "object",
        "description": "Response type for host endpoints.\nIncludes hydrated children (interfaces, ports, services).",
        "required": [
          "id",
          "created_at",
          "updated_at",
          "name",
          "network_id",
          "source",
          "hidden",
          "tags",
          "interfaces",
          "ports",
          "services"
        ],
        "properties": {
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "description": {
            "type": [
              "string",
              "null"
            ]
          },
          "hidden": {
            "type": "boolean"
          },
          "hostname": {
            "type": [
              "string",
              "null"
            ]
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "interfaces": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Interface"
            }
          },
          "name": {
            "type": "string"
          },
          "network_id": {
            "type": "string",
            "format": "uuid"
          },
          "ports": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Port"
            }
          },
          "services": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Service"
            }
          },
          "source": {
            "$ref": "#/components/schemas/EntitySource"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            }
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          },
          "virtualization": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/HostVirtualization"
              }
            ]
          }
        },
        "example": {
          "created_at": "2026-01-15T10:30:00Z",
          "description": "Primary web server",
          "hidden": false,
          "hostname": "web-server-01.local",
          "id": "550e8400-e29b-41d4-a716-446655440003",
          "interfaces": [
            {
              "created_at": "2026-01-15T10:30:00Z",
              "host_id": "550e8400-e29b-41d4-a716-446655440003",
              "id": "550e8400-e29b-41d4-a716-446655440005",
              "ip_address": "192.168.1.100",
              "mac_address": "DE:AD:BE:EF:12:34",
              "name": "eth0",
              "network_id": "550e8400-e29b-41d4-a716-446655440002",
              "subnet_id": "550e8400-e29b-41d4-a716-446655440004",
              "updated_at": "2026-01-15T10:30:00Z"
            }
          ],
          "name": "web-server-01",
          "network_id": "550e8400-e29b-41d4-a716-446655440002",
          "ports": [
            {
              "created_at": "2026-01-15T10:30:00Z",
              "host_id": "550e8400-e29b-41d4-a716-446655440003",
              "id": "550e8400-e29b-41d4-a716-446655440006",
              "network_id": "550e8400-e29b-41d4-a716-446655440002",
              "number": 80,
              "protocol": "Tcp",
              "type": "Http",
              "updated_at": "2026-01-15T10:30:00Z"
            }
          ],
          "services": [],
          "source": {
            "type": "Manual"
          },
          "tags": [],
          "updated_at": "2026-01-15T10:30:00Z",
          "virtualization": null
        }
      },
      "HostVirtualization": {
        "oneOf": [
          {
            "type": "object",
            "title": "Proxmox",
            "required": [
              "details",
              "type"
            ],
            "properties": {
              "details": {
                "$ref": "#/components/schemas/ProxmoxVirtualization"
              },
              "type": {
                "type": "string",
                "enum": [
                  "Proxmox"
                ]
              }
            }
          }
        ],
        "title": "HostVirtualization"
      },
...
      "UpdateHostRequest": {
        "type": "object",
        "description": "Request type for updating a host.\nChildren (interfaces, ports, services) are managed via their own endpoints.",
        "required": [
          "id",
          "name",
          "hidden"
        ],
        "properties": {
          "description": {
            "type": [
              "string",
              "null"
            ]
          },
          "hidden": {
            "type": "boolean"
          },
          "hostname": {
            "type": [
              "string",
              "null"
            ]
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            }
          },
          "virtualization": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/HostVirtualization"
              }
            ]
          }
        }
      },
...
    },
    "securitySchemes": {
      "api_key": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "API key for daemon authentication. Generate keys in the web UI under Settings > API Keys."
      },
      "session": {
        "type": "apiKey",
        "in": "cookie",
        "name": "session"
      }
    }
  }
}